requirements.txt
boto3==1.10.47
botocore==1.13.47
docutils==0.15.2
jmespath==0.9.4
Pillow==7.0.0
python-dateutil==2.8.1
s3transfer==0.2.1
six==1.13.0
urllib3==1.25.7
